Bug 1678733 - Separate out some of the declaration precedence components from @viewport rules r=emilio
Differential Revision: https://phabricator.services.mozilla.com/D97817
