+++ servo/components/style/stylesheets/mod.rs
@@ -4,6 +4,7 @@
 
 //! Style sheets and their CSS rules.
 
+mod cascading_at_rule;
 mod counter_style_rule;
 mod document_rule;
 mod font_face_rule;
